[NCLSUP-794] Improve release script
In this commit dts generator is upgraded to version 3+, there is braking change between dts v2 and v3 where in v3 parts of generator were split into
different projects (plugins). @dtsgenerator/replace-namespace is one of them, we're using it and that's the reason behind added dependency. These plugins
are configured through config file, dtsgen.json in this case.